requirements.txt
streamlit
yfinance
pandas
numpy
transformers
torch
requests
plotly
python-dotenv
finnhub-python
psycopg2-binary
PyYAML
